Testing: move format adaptors to the `"test_format_support.hpp"` header file
