+++ tests/std/tests/P2693R1_text_formatting_thread_id/test.cpp
@@ -18,93 +18,6 @@
 
 using namespace std;
 
-template <class CharT>
-struct FormatFn {
-    template <class... Args>
-    [[nodiscard]] auto operator()(
-        const basic_format_string<CharT, type_identity_t<Args>...> str, Args&&... args) const {
-        return format(str, forward<Args>(args)...);
-    }
-};
-
-template <class CharT>
-struct VFormatFn {
-    template <class... Args>
-    [[nodiscard]] auto operator()(const basic_string_view<CharT> str, Args&&... args) const {
-        if constexpr (same_as<CharT, char>) {
-            return vformat(str, make_format_args(forward<Args>(args)...));
-        } else {
-            return vformat(str, make_wformat_args(forward<Args>(args)...));
-        }
-    }
-};
-
-template <class CharT>
-struct ExpectFormatError {
-private:
-    VFormatFn<CharT> base;
-
-public:
-    template <class... Args>
-    auto operator()(const basic_string_view<CharT> str, Args&&... args) const {
-        try {
-            (void) base(str, forward<Args>(args)...);
-            assert(false && "No exception.");
-        } catch (const format_error&) {
-            return;
-        } catch (...) {
-            assert(false && "Incorrect exception.");
-        }
-    }
-};
-
-template <class CharT>
-struct MoveOnlyFormat {
-private:
-    struct StringInserter {
-        using iterator_category = output_iterator_tag;
-        using difference_type   = ptrdiff_t;
-        using container_type    = basic_string<CharT>;
-
-        StringInserter()                            = default;
-        StringInserter(StringInserter&&)            = default;
-        StringInserter& operator=(StringInserter&&) = default;
-
-        StringInserter(const StringInserter&)            = delete;
-        StringInserter& operator=(const StringInserter&) = delete;
-
-        StringInserter& operator=(const CharT val) {
-            str.push_back(val);
-            return *this;
-        }
-
-        StringInserter& operator*() {
-            return *this;
-        }
-
-        StringInserter& operator++() {
-            return *this;
-        }
-
-        StringInserter operator++(int) {
-            return *this;
-        }
-
-        basic_string<CharT> str;
-    };
-
-public:
-    static_assert(output_iterator<StringInserter, CharT>);
-    static_assert(!copyable<StringInserter>);
-    static_assert(movable<StringInserter>);
-
-    template <class... Args>
-    [[nodiscard]] auto operator()(
-        const basic_format_string<CharT, type_identity_t<Args>...> str, Args&&... args) const {
-        return format_to(StringInserter{}, str, forward<Args>(args)...).str;
-    }
-};
-
 template <class CharT, template <class> class Fmt>
 void check_formatting_of_default_constructed_thread_id() {
     Fmt<CharT> fmt;
